move default query to types
